ibc: making conversions a feature
