+++ ibc/Cargo.toml
@@ -6,7 +6,7 @@ edition = "2018"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-ethereum-types = {version = "0.11"}
+ethereum-types = {version = "0.11", default-features = false}
 celo-types = {path = "../types"}
 rlp = {version = "0.5"}
 base64 = { version = "0.13"}
@@ -15,4 +15,9 @@ schemars = { version = "0.8"}
 anomaly = { version = "0.2", default-features = false }
 thiserror = { version = "1.0", default-features = false }
 
-ibc-proto = {git = "https://github.com/ChorusOne/ibc-rs", default-features = false, branch = "giulio/celochain"}
+#ibc-proto = {git = "https://github.com/ChorusOne/ibc-rs", default-features = false, branch = "giulio/celochain"}
+ibc-proto = {path = "/home/giulio/src/github.com/ChorusOne/ibc-rs/proto", default-features = false, optional = true}
+
+[features]
+#default = ["conversions"]
+conversions = ["ibc-proto"]
