+++ ibc/src/lib.rs
@@ -5,7 +5,11 @@ pub mod header;
 pub mod misbehaviour;
 pub mod proof;
 pub mod state;
+pub mod height;
+#[cfg(features = "conversions")]
+pub mod conversions;
 
+pub use height::Height as Height;
 pub use connection::ConnectionEnd;
 pub use channel::Channel;
 use error::{Error, Kind};
@@ -14,11 +18,3 @@ pub use misbehaviour::Misbehaviour;
 pub use proof::{InnerSpec, LeafOp, MerklePrefix, MerkleRoot, ProofSpec};
 pub use state::{extract_lc_client_state, extract_lc_consensus_state, ClientState, ConsensusState};
 
-pub use ibc_proto::ibc::core::commitment::v1::{MerkleRoot as IBCMerkleRoot, MerklePrefix as IBCMerklePrefix};
-pub use ibc_proto::ibc::lightclients::wasm::v1::ClientState as IBCClientState;
-pub use ibc_proto::ibc::lightclients::wasm::v1::ConsensusState as IBCConsensusState;
-pub use ibc_proto::ibc::lightclients::wasm::v1::Header as IBCHeader;
-pub use ibc_proto::ibc::lightclients::wasm::v1::Misbehaviour as IBCMisbehaviour;
-pub use ibc_proto::ics23::InnerSpec as ICSInnerSpec;
-pub use ibc_proto::ics23::LeafOp as ICSLeafOp;
-pub use ibc_proto::ics23::ProofSpec as ICSProofSpec;
